refactor(global): 예외 처리 로직 global로 이동
